Modify output file to be newick str with comments
I modified the treeBreaker so that the output file now is the
newick string with comments. The comments are in [] and can have
three sections. "index" is the index of the node that is used to return
b. "pheno" is the phenotype of the leaf nodes. "posterior" is the
posterior probability that has been inferred by the MCMC.