Use unrooted tree instead of rooted tree
I have modified the algorithm such that it uses unrooted trees.
This was because the two branches under the root node are just one
branch and having a change point on either one or both of those
branches result in the same divisions on the tree.

Now I only allow change points on the longer branch.