Add model comparison
I added the rjmcmc to calculate the bayes factor for the two models.
In model 0, lambda is zero and in model 1 lambda is non-zero. In case of
lambda == zero, there are no change points on the tree. in case of
lambda non-zero, there might be change points on the tree.