Add hla footprinting interesting results
I have added the results where the BF is bigger than 2.
The sample with BF = inf is 1-1000.newick_col_91_phenos.csv.output sample
where the tree of the first 1000 bases is associated with B57 hla type.
When we look at the tree there is a small clade of 10-15 isolates where
most of the hosts have hla B57. This is really surprising. Either these
are a group of people who are closely related and have transmitted their hiv to
each other. The other explanation is that there is a region in the first 1000 bases
that is affected by B57. We need to know more about the patients.